flask-backend/flask-backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
